Add files via upload
Deep learning techniques are employed for the purpose of differentiating chest X-ray images among patients with Covid-19, viral pneumonia, bacterial pneumonia, and those categorized as normal. This involves the utilization of advanced convolutional neural networks (CNNs) as a powerful tool in the field of medical image analysis. The model is trained on a diverse dataset comprising labeled images from each category, and through this process, it learns to distinguish pathological conditions, contributing to the potential improvement of diagnostic accuracy in healthcare.